feat: ide awareness during onboarding
