+++ components/dashboard/src/workspaces/Workspaces.tsx
@@ -13,8 +13,11 @@ import { WorkspaceEntry } from "./WorkspaceEntry";
 import { getGitpodService } from "../service/service";
 import { ItemsList } from "../components/ItemsList";
 import { TeamsContext } from "../teams/teams-context";
+import { UserContext } from "../user-context";
+import { User } from "@gitpod/gitpod-protocol";
 import { useLocation } from "react-router";
 import { StartWorkspaceModalContext, StartWorkspaceModalKeyBinding } from "./start-workspace-modal-context";
+import SelectIDEModal from "../settings/SelectIDEModal";
 
 export interface WorkspacesProps {}
 
@@ -27,6 +30,7 @@ export interface WorkspacesState {
 export default function () {
     const location = useLocation();
 
+    const { user } = useContext(UserContext);
     const { teams } = useContext(TeamsContext);
     const [activeWorkspaces, setActiveWorkspaces] = useState<WorkspaceInfo[]>([]);
     const [inactiveWorkspaces, setInactiveWorkspaces] = useState<WorkspaceInfo[]>([]);
@@ -40,9 +44,13 @@ export default function () {
         })();
     }, [teams, location]);
 
+    const isOnboardingUser = user && User.isOnboardingUser(user);
+
     return (
         <>
             <Header title="Workspaces" subtitle="Manage recent and stopped workspaces." />
+
+            {isOnboardingUser && <SelectIDEModal />}
 
             {workspaceModel?.initialized &&
                 (activeWorkspaces.length > 0 || inactiveWorkspaces.length > 0 || workspaceModel.searchTerm ? (
