+++ components/gitpod-db/src/typeorm/user-db-impl.ts
@@ -101,7 +101,7 @@ export class TypeORMUserDBImpl implements UserDB {
             creationDate: new Date().toISOString(),
             identities: [],
             additionalData: {
-                ideSettings: { defaultIde: "code" },
+                ideSettings: {},
                 emailNotificationSettings: {
                     allowsChangelogMail: true,
                     allowsDevXMail: true,
